requirements.txt
celery[redis]
flask
numpy
scipy
scikit-learn
python-dateutil